added sources file
btw for some reason 1.7.10 doesn't output a source file for me :((( but that's fine

note: you might have noticed this one is cloudmc-1.0-sources.jar as opposed to cloud-1.8-sources.jar. this is because the build outputs files named cloudmc-1.0, not cloud-(minecraft version). the files get renamed before getting uploaded, both here and on the official cloud repo.